backend/requirements.txt
django 
djangorestframework 
psycopg2-binary 
djangorestframework-simplejwt 
pandas 
nltk 
sklearn  
scipy 
pytz 
xgboost==0.90

